+++ control/PID/PID.h
@@ -8,13 +8,15 @@
 #ifndef PID_H
 #define PID_H
 
+#include <stdint.h>
+
 typedef struct {
 	double P_gain;
 	double I_gain;
 	double D_gain;
 	double e_prev;
 	double e_cumulative;
-	int t_prev;
+	uint64_t t_prev;
 } PID_controller;
 
 /**@brief Generate an actuator command.
@@ -29,7 +31,7 @@ typedef struct {
 double PID_command(
 	double target,
 	double state,
-	int t_curr,
+	uint64_t t_curr,
 	PID_controller *controller
 );
 
@@ -48,7 +50,7 @@ double PID_command(
 void PID_init(
 	double target_init,
 	double state_init,
-	int t_init,
+	uint64_t t_init,
 	double P_gain,
 	double I_gain,
 	double D_gain,
